requirements.txt
telethon
openpyxl
ollama
python-dotenv
